chore(context-engine): satisfy max-lines and assertion-safety ratchets
Extract the runtime quarantine state group from registry.ts into
quarantine-state.ts (registry keeps the public re-export) so the file
stays under the max-lines threshold. Replace the two non-const type
assertions flagged by the assertion-safety ratchet with type-safe
constructs: in-operator narrowing for the maintenance error marker and a
direct NodeJS.Process annotation for the abort-signal fallback. Adjust
the grandfathered baseline for context-engine-maintenance.ts to record
the already-removed assertion.
