+++ config/assertion-safety-baseline.txt
@@ -1808,7 +1808,7 @@ src/agents/embedded-agent-runner/compact.queued.ts	5
 src/agents/embedded-agent-runner/compaction-diagnostics.ts	5
 src/agents/embedded-agent-runner/compaction-hooks.ts	1
 src/agents/embedded-agent-runner/compaction-session-agent.ts	7
-src/agents/embedded-agent-runner/context-engine-maintenance.ts	3
+src/agents/embedded-agent-runner/context-engine-maintenance.ts	2
 src/agents/embedded-agent-runner/delivery-evidence.ts	12
 src/agents/embedded-agent-runner/direct-compaction-preparation.ts	1
 src/agents/embedded-agent-runner/empty-assistant-turn.ts	1
